Clean up buttons, separate out properties, add xml docs where needed
