+++ FluentAvalonia/UI/Controls/Button/DropDownButton.cs
@@ -1,10 +1,11 @@
 ﻿using Avalonia.Styling;
 using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace FluentAvalonia.UI.Controls
 {
+    /// <summary>
+    /// Represents a button that includes a chevron to indicate a menu can be opened.
+    /// </summary>
     public class DropDownButton : Button, IStyleable
     {
         Type IStyleable.StyleKey => typeof(DropDownButton);
