+++ FluentAvalonia/UI/Controls/SplitButton/SplitButtonClickEventArgs.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace FluentAvalonia.UI.Controls
+{
+	public class SplitButtonClickEventArgs : EventArgs 
+	{ 
+
+	}
+}
