+++ FluentAvalonia/UI/Controls/SplitButton/ToggleSplitButtonIsCheckedChangedEventArgs.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace FluentAvalonia.UI.Controls
+{
+	public class ToggleSplitButtonIsCheckedChangedEventArgs : EventArgs
+	{
+
+	}
+}
